requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
PyMySQL==1.1.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1
pydantic==2.8.2